add create_spatial dataset
